LayoutTests/imported/w3c/web-platform-tests/selection/extend-40-expected.txt


FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 0 [paras[0].firstChild, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 0 [paras[0].firstChild, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 1 [paras[0].firstChild, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 1 [paras[0].firstChild, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 2 [paras[0].firstChild, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 2 [paras[0].firstChild, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 3 [paras[0].firstChild, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 3 [paras[0].firstChild, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 4 [paras[0].firstChild, 8] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 4 [paras[0].firstChild, 8] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 5 [paras[0].firstChild, 9] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 5 [paras[0].firstChild, 9] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 6 [paras[0].firstChild, 10] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 6 [paras[0].firstChild, 10] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 7 [paras[0].firstChild, 65535] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 7 [paras[0].firstChild, 65535] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 8 [paras[1].firstChild, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 8 [paras[1].firstChild, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 9 [paras[1].firstChild, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 9 [paras[1].firstChild, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 10 [paras[1].firstChild, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 10 [paras[1].firstChild, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 11 [paras[1].firstChild, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 11 [paras[1].firstChild, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 12 [paras[1].firstChild, 8] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 12 [paras[1].firstChild, 8] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 13 [paras[1].firstChild, 9] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 13 [paras[1].firstChild, 9] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 14 [paras[1].firstChild, 10] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 14 [paras[1].firstChild, 10] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 15 [paras[1].firstChild, 65535] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 15 [paras[1].firstChild, 65535] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 16 [detachedPara1.firstChild, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 16 [detachedPara1.firstChild, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 17 [detachedPara1.firstChild, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 17 [detachedPara1.firstChild, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 18 [detachedPara1.firstChild, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 18 [detachedPara1.firstChild, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 19 [detachedPara1.firstChild, 9] The index is not in the allowed range.
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 19 [detachedPara1.firstChild, 9] The index is not in the allowed range.
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 20 [foreignPara1.firstChild, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 20 [foreignPara1.firstChild, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 21 [foreignPara1.firstChild, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 21 [foreignPara1.firstChild, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 22 [foreignPara1.firstChild, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 22 [foreignPara1.firstChild, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 23 [foreignPara1.firstChild, 9] The index is not in the allowed range.
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 23 [foreignPara1.firstChild, 9] The index is not in the allowed range.
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 24 [document.documentElement, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 24 [document.documentElement, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 25 [document.documentElement, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 25 [document.documentElement, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 26 [document.documentElement, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 26 [document.documentElement, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 27 [document.documentElement, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 27 [document.documentElement, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 28 [document.documentElement, 7] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 28 [document.documentElement, 7] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 29 [document.head, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 29 [document.head, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 30 [document.body, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 30 [document.body, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 31 [foreignDoc.documentElement, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 31 [foreignDoc.documentElement, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 32 [foreignDoc.documentElement, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 32 [foreignDoc.documentElement, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 33 [foreignDoc.head, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 33 [foreignDoc.head, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 34 [foreignDoc.body, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 34 [foreignDoc.body, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 35 [paras[0], 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 35 [paras[0], 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 36 [paras[0], 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 36 [paras[0], 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 37 [paras[0], 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 37 [paras[0], 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 38 [paras[1], 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 38 [paras[1], 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 39 [paras[1], 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 39 [paras[1], 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 40 [paras[1], 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 40 [paras[1], 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 41 [detachedPara1, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 41 [detachedPara1, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 42 [detachedPara1, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 42 [detachedPara1, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 43 [testDiv, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 43 [testDiv, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 44 [testDiv, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 44 [testDiv, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 45 [document, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 45 [document, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 46 [document, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 46 [document, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 47 [document, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 47 [document, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 48 [document, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 48 [document, 2] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 49 [document, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 49 [document, 3] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 50 [comment, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 50 [comment, -1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 51 [comment, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 51 [comment, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 52 [comment, 4] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 52 [comment, 4] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 53 [comment, 96] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 53 [comment, 96] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 54 [foreignDoc, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 54 [foreignDoc, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 55 [foreignDoc, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 55 [foreignDoc, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 56 [foreignComment, 2] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 56 [foreignComment, 2] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 57 [foreignTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 57 [foreignTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 58 [foreignTextNode, 36] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 58 [foreignTextNode, 36] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 59 [xmlDoc, -1] The index is not in the allowed range.
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 59 [xmlDoc, -1] The index is not in the allowed range.
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 60 [xmlDoc, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 60 [xmlDoc, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 61 [xmlDoc, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 61 [xmlDoc, 1] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 62 [xmlDoc, 5] The index is not in the allowed range.
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 62 [xmlDoc, 5] The index is not in the allowed range.
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 63 [xmlComment, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 63 [xmlComment, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 64 [xmlComment, 4] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 64 [xmlComment, 4] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 65 [processingInstruction, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 65 [processingInstruction, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 66 [processingInstruction, 5] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 66 [processingInstruction, 5] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 67 [processingInstruction, 9] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 67 [processingInstruction, 9] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 68 [detachedTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 68 [detachedTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 69 [detachedTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 69 [detachedTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 70 [detachedForeignTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 70 [detachedForeignTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 71 [detachedForeignTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 71 [detachedForeignTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 72 [detachedXmlTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 72 [detachedXmlTextNode, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 73 [detachedXmlTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 73 [detachedXmlTextNode, 8] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 74 [detachedProcessingInstruction, 12] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 74 [detachedProcessingInstruction, 12] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 75 [detachedComment, 3] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 75 [detachedComment, 3] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 76 [detachedComment, 5] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 76 [detachedComment, 5] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 77 [detachedForeignComment, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 77 [detachedForeignComment, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 78 [detachedForeignComment, 4] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 78 [detachedForeignComment, 4] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 79 [detachedXmlComment, 2] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 79 [detachedXmlComment, 2] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 80 [docfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 80 [docfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 81 [foreignDocfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 81 [foreignDocfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 82 [xmlDocfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 82 [xmlDocfrag, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 83 [doctype, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 83 [doctype, 0] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 84 [doctype, -17] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 84 [doctype, -17] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 85 [doctype, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 85 [doctype, 1] assert_equals: Sanity check: startContainer must be correct expected Element node <p id="d" style="display:none">Yzabcdef</p> but got Text node "Qrstuvwx"
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 86 [foreignDoctype, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 86 [foreignDoctype, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() forwards with range 40 [paras[3], 1, comment, 8] and point 87 [xmlDoctype, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
FAIL extend() backwards with range 40 [paras[3], 1, comment, 8] and point 87 [xmlDoctype, 0] assert_equals: The operation should not replace a registered Range expected object "" but got object ""
